Isolated some functions in own file.
